Splited bundleOptions tests to two platforms
