+++ fixtures/react_native_with_haul_0_60x_ts/tsconfig.json
@@ -7,7 +7,7 @@
     "lib": ["es2016"],
     "module": "commonjs",
     "moduleResolution": "node",
-    "jsx": "react-native",
+    "jsx": "react",
     "resolveJsonModule": true
   }
 }
